Refactor infinispan-grpc-kafka sceanrio due to a native error
