+++ messaging/infinispan-grpc-kafka/src/test/java/io/quarkus/qe/messaging/infinispan/InfinispanKafkaSslIT.java
@@ -3,35 +3,29 @@ package io.quarkus.qe.messaging.infinispan;
 import static io.restassured.RestAssured.get;
 import static io.restassured.RestAssured.given;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.Matchers.containsString;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.apache.http.HttpStatus;
 import org.hamcrest.core.StringContains;
 import org.junit.jupiter.api.Test;
 
-import io.quarkus.qe.messaging.infinispan.books.Book;
 import io.quarkus.test.bootstrap.InfinispanService;
 import io.quarkus.test.bootstrap.KafkaService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.scenarios.annotations.DisabledOnNative;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.KafkaContainer;
 import io.quarkus.test.services.QuarkusApplication;
 import io.quarkus.test.services.containers.model.KafkaProtocol;
 import io.quarkus.test.services.containers.model.KafkaVendor;
-import io.restassured.http.ContentType;
 
 @QuarkusScenario
-public class BookServiceIT {
+@DisabledOnNative(reason = "https://github.com/quarkusio/quarkus/issues/25098")
+public class InfinispanKafkaSslIT {
 
     /**
      * We can't rename this file to use the default SSL settings part of KafkaService.
      */
     private static final String TRUSTSTORE_FILE = "strimzi-server-ssl-truststore.p12";
-    private static final String BOOK_TITLE = "testBook";
-    private static final Book BOOK = new Book(BOOK_TITLE, "description", 2011);
 
     @Container(image = "${infinispan.image}", expectedLog = "${infinispan.expected-log}", port = 11222)
     static final InfinispanService infinispan = new InfinispanService()
@@ -40,9 +34,6 @@ public class BookServiceIT {
 
     @KafkaContainer(vendor = KafkaVendor.STRIMZI, protocol = KafkaProtocol.SSL, kafkaConfigResources = TRUSTSTORE_FILE)
     static final KafkaService kafkassl = new KafkaService();
-
-    @KafkaContainer(vendor = KafkaVendor.STRIMZI, protocol = KafkaProtocol.SASL)
-    static final KafkaService kafkasasl = new KafkaService();
 
     @QuarkusApplication
     static final RestService app = new RestService()
@@ -60,82 +51,7 @@ public class BookServiceIT {
             .withProperty("kafka-streams.state.dir", "target")
             .withProperty("quarkus.kafka-streams.ssl.endpoint-identification-algorithm", "")
             .withProperty("quarkus.kafka-streams.bootstrap-servers", kafkassl::getBootstrapUrl)
-            .withProperty("kafka-client-ssl.bootstrap.servers", kafkassl::getBootstrapUrl)
-            .withProperty("kafka-client-sasl.bootstrap.servers", kafkasasl::getBootstrapUrl);
-
-    @Test
-    public void testBookResource() {
-        given()
-                .contentType(ContentType.JSON)
-                .body(BOOK)
-                .when().post("/book/add")
-                .then().statusCode(HttpStatus.SC_NO_CONTENT);
-
-        Book actual = given()
-                .accept(ContentType.JSON)
-                .when().get("/book/" + BOOK_TITLE)
-                .as(Book.class);
-
-        assertEquals(BOOK, actual);
-    }
-
-    @Test
-    public void testBookResourceShouldValidateBook() {
-        given()
-                .contentType(ContentType.JSON)
-                .body(new Book())
-                .when().post("/book/add")
-                .then().statusCode(HttpStatus.SC_BAD_REQUEST)
-                .body(containsString("Title cannot be blank"));
-    }
-
-    @Test
-    public void testBlockingGreetingResource() {
-        given()
-                .when().get("/hello/blocking/neo")
-                .then().statusCode(200)
-                .body(is("Hello neo"));
-    }
-
-    @Test
-    public void testMutinyGreetingResource() {
-        given()
-                .when().get("/hello/mutiny/neo")
-                .then().statusCode(200)
-                .body(is("Hello neo"));
-    }
-
-    @Test
-    public void testInfinispanEndpoint() {
-        given()
-                .when().get("/infinispan")
-                .then()
-                .statusCode(200)
-                .body(is("Hello World, Infinispan is up!"));
-    }
-
-    @Test
-    void testKafkaClientSASL() {
-        await().untilAsserted(() -> {
-            given()
-                    .queryParam("key", "my-key")
-                    .queryParam("value", "my-value")
-                    .when()
-                    .post("/kafka/sasl")
-                    .then()
-                    .statusCode(200);
-
-            get("/kafka/sasl")
-                    .then()
-                    .statusCode(200)
-                    .body(StringContains.containsString("my-key-my-value"));
-        });
-
-        get("/kafka/sasl/topics")
-                .then()
-                .statusCode(200)
-                .body(StringContains.containsString("hello"));
-    }
+            .withProperty("kafka-client-ssl.bootstrap.servers", kafkassl::getBootstrapUrl);
 
     @Test
     void testKafkaClientSSL() {
